requirements.txt
Django==4.2.24
djangorestframework==3.16.1
djangorestframework-api-response==0.1.0
django-redis==6.0.0
django-environ==0.12.0
pillow==11.3.0
django-ckeditor==6.7.3

channels==4.3.1
channels-redis==4.3.0

uvicorn==0.35.0

whitenoise==6.10.0
psycopg2==2.9.10
celery==5.5.3
django-celery-results==2.6.0
django-celery-beat==2.8.1

Faker==37.8.0



